+++ extension/core/ELanguage/AutoLanguage.js
@@ -1,0 +1,5 @@
+
+export default function AutoLang() {
+  var webLanguage = navigator.language.split('-')[0]
+  return webLanguage
+}